fbs-print-agent/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0
pywin32>=306
pystray>=0.19.0
Pillow>=10.0
img2pdf>=0.5.0
